vAPI: Move u_auth into own file
